COMTESS version - singleinds: debugged: csal has 0/1 effect
git-svn-id: file:///X/DropKatrinK/SVNRepository/Grassland/COMTESS/branches/singleinds@586 0913833f-c9c6-fa4f-9488-fa0ee326e6e0

+++ CWaterGridEnvir.h
@@ -191,7 +191,7 @@ struct SWaterTraits: public SPftTraits   //plant functional traits
    /// else it's efficiency is reduced by this value for all WL
    ///
    double assimAnoxWL;
-   int saltTol; ///< salt tolerance of species (Ellenberg salt value)
+   double saltTol; ///< salt tolerance of species (Ellenberg salt value)
    double saltTolCosts(); ///< salt tolerance costs
    double saltTolEffect(double salinity); ///< salt tolerance effect
